+++ hypers/_tools/__init__.py
@@ -1,6 +1,0 @@
-from hypers._tools._smoothen import _data_smoothen
-from hypers._tools._types import PreprocessType, ClusterType, DecomposeType, MixtureType
-from hypers._tools._update import _data_mean, _data_checks, _data_access
-
-__all__ = ['_data_smoothen', '_data_mean', '_data_checks', '_data_access', 'PreprocessType',
-           'ClusterType', 'DecomposeType', 'MixtureType']