+++ hypers/learning/__init__.py
@@ -1,0 +1,4 @@
+# from hypers.learning._abundance import _ucls
+# from hypers.learning._mixture import _data_mixture
+
+# __all__ = ['_kmeans', '_vca', '_ucls', '_data_mixture', '_pca', '_ica', '_nmf']