+++ hypers/signal/__init__.py
@@ -1,0 +1,3 @@
+from hypers.signal.core import map_units
+
+__all__ = ['map_units']